reorganize functions related to chat cards
