+++ module/chat.ts
@@ -1,6 +1,70 @@
 import { SR5Actor } from './actor/SR5Actor';
 import { SR5Item } from './item/SR5Item';
 import Template from './template';
+import ModList = Shadowrun.ModList;
+import BaseValuePair = Shadowrun.BaseValuePair;
+import DamageData = Shadowrun.DamageData;
+import AttackData = Shadowrun.AttackData;
+import LabelField = Shadowrun.LabelField;
+
+export type TemplateData = {
+    header: {
+        name: string;
+        img: string;
+    };
+    tokenId?: string;
+    dice?: Die[];
+    parts?: ModList<number>;
+    limit?: BaseValuePair<number> & LabelField;
+    testName?: string;
+    actor?: SR5Actor;
+    item?: SR5Item;
+    attack?: AttackData;
+    incomingAttack?: AttackData;
+    incomingDrain?: LabelField & {
+        value: number;
+    };
+    soak?: DamageData;
+    tests?: {
+        label: string;
+        type: string;
+    }[];
+    description?: object;
+    previewTemplate?: boolean;
+};
+
+export const createChatData = async (templateData: TemplateData, roll?) => {
+    const template = `systems/shadowrun5e/templates/rolls/roll-card.html`;
+    const html = await renderTemplate(template, templateData);
+    const actor = templateData.actor;
+
+    const chatData = {
+        user: game.user._id,
+        type: CONST.CHAT_MESSAGE_TYPES.ROLL,
+        content: html,
+        roll,
+        speaker: {
+            actor: actor?._id,
+            token: actor?.token,
+            alias: actor?.name,
+        },
+        flags: {
+            shadowrun5e: {
+                customRoll: true,
+            },
+        },
+    };
+    if (roll) {
+        chatData['sound'] = CONFIG.sounds.dice;
+    }
+    const rollMode = game.settings.get('core', 'rollMode');
+
+    if (['gmroll', 'blindroll'].includes(rollMode))
+        chatData['whisper'] = ChatMessage.getWhisperIDs('GM');
+    if (rollMode === 'blindroll') chatData['blind'] = true;
+
+    return chatData;
+};
 
 export const addChatMessageContextOptions = function (html, options) {
     const canRoll = (li) => {
@@ -30,7 +94,7 @@ export const addChatMessageContextOptions = function (html, options) {
     return options;
 };
 
-export const addRollListeners = (app, html) => {
+export const addRollListeners = (app: ChatMessage, html) => {
     if (!app.getFlag('shadowrun5e', 'customRoll')) return;
     html.on('click', '.test-roll', async (event) => {
         event.preventDefault();
@@ -42,7 +106,7 @@ export const addRollListeners = (app, html) => {
                 const html = await renderTemplate(template, roll.templateData);
                 const data = {};
                 data['content'] = html;
-                app.update(data);
+                await app.update(data);
             }
         }
     });
